apps/ai-worker/pyproject.toml
[project]
name = "persona-mirror-ai-worker"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "anthropic>=0.86.0",
    "langgraph>=1.1.3",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]

